Refactor base abstraction, extract detection methods into concern
